# services/auth-service/contracts/openapi/auth.yaml
openapi: 3.0.3
info:
  title: PRAHARI Auth Service API
  description: API contract specifications for User authentication, Registration, and JWT token management.
  version: 1.0.0
servers:
  - url: http://localhost:8001
    description: Local Dev Server
paths:
...
  /api/v1/auth/refresh:
    post:
      summary: Refresh active JWT access tokens
      operationId: refreshToken
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshRequest'
      responses:
        '200':
          description: Tokens refreshed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '401':
          $ref: '#/components/responses/401Error'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    RefreshRequest:
      type: object
      required:
        - refresh_token
      properties:
        refresh_token:
          type: string
    TokenResponse:
      type: object
      properties:
        access_token:
          type: string
        id_token:
          type: string
        refresh_token:
          type: string
        expires_in:
          type: integer
...
    ErrorEnvelope:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
            details:
              type: array
              items:
                type: object
  responses:
...
    401Error:
      description: Unauthenticated
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
...